requirements.txt
streamlit
streamlit-shap 
pandas>=2.1.3
numpy>=1.26.2
matplotlib>=3.2.2
seaborn==0.13.0
plotly>=5.18.0
requests==2.31.0
Pillow==10.0.1
scikit_learn==1.3.2
lightgbm==4.1.0
shap>=0.44
requests
fastapi
pillow
uvicorn
gunicorn
pydantic